feat(manifest): `loader.Load(...)` includes unselected environments as not enabled environments with `Enabled: false`
In order to be able to check environment and group overrides in `project.LoadProject(...)` they need not to be removed in `loader.Load(...)`. Instead, this function marks them as not enabled, with `Enabled: false`.

+++ cmd/monaco/purge/purge.go
@@ -63,6 +63,9 @@ func purge(ctx context.Context, fs afero.Fs, deploymentManifestPath string, envi
 func purgeConfigs(ctx context.Context, environments []manifest.EnvironmentDefinition, apis api.APIs) error {
 
 	for _, env := range environments {
+		if !env.Enabled {
+			continue
+		}
 		err := purgeForEnvironment(ctx, env, apis)
 		if err != nil {
 			return err
